tests/wpt/web-platform-tests/orientation-event/t023-manual.html

<!DOCTYPE html>
<html>
  <head>
    <title>User agents must also provide an event handler IDL attribute HTML5 named ondevicemotion on the window object</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="author" title="Mosquito FP7">
    <link rel="reviewer author" title="Zhiqiang Zhang" href="mailto:zhiqiang.zhang@intel.com"> <!-- 2013-10-14 -->
  </head>
  <body>
    <p>Move the device to run the test.</p>
    <div id="log"></div>
    <script>
      var t1 = async_test("User agents provide an event handler IDL attribute HTML5 named ondevicemotion");
      var t2 = async_test("The type of this event handler must be 'DeviceMotionEvent'");
      var run = false;
      window.ondevicemotion = function(e) {
        if (!run) {
          run = true;
          t1.step(function() {
            assert_equals(e.type, "devicemotion");
          });
          t1.done();
          t2.step(function() {
            assert_true(e instanceof "DeviceMotionEvent");
          });
          t2.done();
        }
      };
    </script>
  </body>
</html>
